fix: Windows build errors (libc::kill, c_long cast, XactEvent::Type)
- shmem.rs: wrap libc::kill/pid_t/SIGHUP in #[cfg(unix)] — these
  POSIX signals do not exist on Windows
- sparql/cursor.rs: cast page_size (i64) to libc::c_long before
  calling SpiCursor::fetch(); on Windows c_long is i32, not i64
- lib.rs: change xact_callback_c event parameter from u32 to
  pgrx::pg_sys::XactEvent::Type and use named constants
  (XACT_EVENT_ABORT etc.) — on Windows pgrx bindgen emits c_int
  (i32) for XactEvent::Type rather than c_uint (u32)

+++ src/sparql/cursor.rs
@@ -136,7 +136,7 @@ impl CursorIter {
                 .unwrap_or_else(|e| pgrx::error!("CursorIter: find_cursor failed: {e}"));
 
             let table = cursor
-                .fetch(page_size)
+                .fetch(page_size as libc::c_long)
                 .unwrap_or_else(|e| pgrx::error!("CursorIter: fetch failed: {e}"));
 
             // Collect raw values and all dictionary IDs in one pass.
@@ -416,7 +416,7 @@ impl Iterator for ConstructCursorIter {
                 .unwrap_or_else(|e| pgrx::error!("ConstructCursorIter: find_cursor failed: {e}"));
 
             let table = cursor
-                .fetch(page_size)
+                .fetch(page_size as libc::c_long)
                 .unwrap_or_else(|e| pgrx::error!("ConstructCursorIter: fetch failed: {e}"));
 
             // Collect raw variable bindings (one i64 per variable per row).
